+++ packages/steiner/src/helpers/formHelper.js
@@ -1,11 +1,13 @@
 import { createFormAction } from './reduxFormSaga';
 
-export function createSubmit(actionTypes) {
+const identity = data => data;
+
+export function createSubmit(actionTypes, dataTransformer = identity) {
     return function submit(data, dispatch) {
         const action = data.id
             ? createFormAction(actionTypes.update, [actionTypes.updateSuccess, actionTypes.updateFail])
             : createFormAction(actionTypes.create, [actionTypes.createSuccess, actionTypes.createFail]);
 
-        return action(data, dispatch);
+        return action(dataTransformer(data), dispatch);
     }
 }