+++ packages/steiner-cli/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0 (October 3, 2016)
+
+- Updated edit template
+
 ## 0.1.0 (September 30, 2016)
 
 First release!