+++ packages/steiner/CHANGELOG.md
@@ -1,6 +1,9 @@
-## HEAD
+## 0.2.0 (October 3, 2016)
 - Fixed some minor bugs
 - Improved error handling on form submit
+- Added support to query string endpoints on `createApi`
+- Added support to transform data before submit on `createSubmit` helper
+- `connectItem` does not pass `initialValues` props anymore
 
 ## 0.1.0 (September 30, 2016)